Bailing on SBE because test server is easier to deploy with QuickFIX which doesn't support SBE
