+++ .gitignore
@@ -72,3 +72,6 @@ Thumbs.db
 # Ignore debugging symbols
 *.dSYM
 
+#Ignore notetaking files
+change-notes.txt
+
