+++ include/FIXMessageBuilder.hpp
@@ -1,22 +1,9 @@
 #pragma once
-#include <cinttypes>
-#include <cstdio>
-#include <iomanip>
-#include <iostream>
-#include <string>
-
-#include "MessageHeader.h"
-#include "NewOrderSingle.h"
 
 class FIXMessageBuilder {
  public:
   FIXMessageBuilder(/* args */);
-  auto EncodeHeader(sbe::MessageHeader& header, sbe::NewOrderSingle& new_order,
-                    char* buffer, std::uint64_t offset,
-                    std::uint64_t buffer_len) -> std::size_t;
-  auto EncodeMessage(sbe::NewOrderSingle& new_order, char* buffer,
-                     std::uint64_t offset, std::uint64_t buffer_len)
-      -> std::size_t;
+  ~FIXMessageBuilder();
 
  private:
   /* data */
